- migrated to existing search composable - got rid of own searchbar
