+++ app/src/main/java/es/jvbabi/vplanplus/ui/screens/Screen.kt
@@ -2,8 +2,7 @@ package es.jvbabi.vplanplus.ui.screens
 
 sealed class Screen(val route: String) {
     data object HomeScreen: Screen("home_screen")
-    data object SearchScreen: Screen("search_screen")
-    data object SearchAvailableRoomScreen: Screen("search_screen/available_room")
+    data object SearchAvailableRoomScreen: Screen("home_screen/available_room")
 
     data object Onboarding: Screen("onboarding")
     data object OnboardingWelcomeScreen: Screen("onboarding/welcome_screen")
